[sw] Fix build after moving abs_mmio.
