+++ sw/device/lib/base/abs_mmio.c
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "sw/device/silicon_creator/lib/base/abs_mmio.h"
+#include "sw/device/lib/base/abs_mmio.h"
 
 // `extern` declarations to give the inline functions in the corresponding
 // header a link location.
