+++ sw/device/silicon_creator/lib/base/sec_mmio.c
@@ -6,9 +6,9 @@
 
 #include <assert.h>
 
+#include "sw/device/lib/base/abs_mmio.h"
 #include "sw/device/lib/base/hardened.h"
 #include "sw/device/lib/base/macros.h"
-#include "sw/device/silicon_creator/lib/base/abs_mmio.h"
 
 // The context is declared as weak so that the mask ROM and ROM_EXT may
 // override its location.
