+++ sw/device/silicon_creator/lib/drivers/meson.build
@@ -26,7 +26,7 @@ sw_silicon_creator_lib_driver_hmac = declare_dependency(
       'hmac.c',
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
     ],
   ),
 )
@@ -73,7 +73,7 @@ sw_silicon_creator_lib_driver_keymgr = declare_dependency(
       'keymgr.c',
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
       sw_silicon_creator_lib_base_sec_mmio,
     ],
   ),
@@ -133,7 +133,7 @@ sw_silicon_creator_lib_driver_uart = declare_dependency(
     ],
     dependencies: [
       sw_lib_mmio,
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
     ],
   ),
 )
@@ -211,7 +211,7 @@ sw_silicon_creator_lib_driver_retention_sram = declare_dependency(
       'retention_sram.c',
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
     ],
   ),
 )
@@ -259,7 +259,7 @@ sw_silicon_creator_lib_driver_rnd = declare_dependency(
       'rnd.c',
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
       sw_silicon_creator_lib_driver_otp,
     ],
   ),
@@ -289,7 +289,7 @@ sw_silicon_creator_lib_driver_rstmgr = declare_dependency(
       'rstmgr.c',
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
     ],
   ),
 )
@@ -321,7 +321,7 @@ sw_silicon_creator_lib_driver_rstmgr_functest = declare_dependency(
       'rstmgr_functest.c'
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
       sw_silicon_creator_lib_driver_rstmgr,
       sw_silicon_creator_lib_driver_retention_sram,
     ],
@@ -376,7 +376,7 @@ sw_silicon_creator_lib_driver_alert_functest = declare_dependency(
       'alert_functest.c'
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
       sw_silicon_creator_lib_base_sec_mmio,
       sw_silicon_creator_lib_driver_alert,
       sw_silicon_creator_lib_driver_rstmgr,
@@ -398,7 +398,7 @@ sw_silicon_creator_lib_driver_pinmux = declare_dependency(
       'pinmux.c',
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
     ],
   ),
 )
@@ -432,7 +432,7 @@ sw_silicon_creator_lib_driver_watchdog = declare_dependency(
     ],
     dependencies: [
       sw_lib_mmio,
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
     ],
   ),
 )
@@ -465,7 +465,7 @@ sw_silicon_creator_lib_driver_otbn = declare_dependency(
       'otbn.c',
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
     ],
   ),
 )
@@ -518,7 +518,7 @@ sw_silicon_creator_lib_driver_flash_ctrl = declare_dependency(
       'flash_ctrl.c',
     ],
     dependencies: [
-      sw_silicon_creator_lib_base_abs_mmio,
+      sw_lib_abs_mmio,
       sw_silicon_creator_lib_base_sec_mmio,
       sw_silicon_creator_lib_driver_otp,
       sw_lib_hardened,
