+++ sw/device/silicon_creator/lib/drivers/pinmux.c
@@ -4,8 +4,8 @@
 
 #include "sw/device/silicon_creator/lib/drivers/pinmux.h"
 
+#include "sw/device/lib/base/abs_mmio.h"
 #include "sw/device/lib/base/macros.h"
-#include "sw/device/silicon_creator/lib/base/abs_mmio.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
 #include "pinmux_regs.h"  // Generated.
