+++ sw/device/silicon_creator/lib/drivers/rnd.c
@@ -4,9 +4,9 @@
 
 #include "sw/device/silicon_creator/lib/drivers/rnd.h"
 
+#include "sw/device/lib/base/abs_mmio.h"
 #include "sw/device/lib/base/csr.h"
 #include "sw/device/lib/base/hardened.h"
-#include "sw/device/silicon_creator/lib/base/abs_mmio.h"
 #include "sw/device/silicon_creator/lib/drivers/otp.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
