+++ sw/device/silicon_creator/lib/drivers/rstmgr.c
@@ -14,7 +14,7 @@
 #include "sw/device/lib/runtime/hart.h"
 #endif
 
-#include "sw/device/silicon_creator/lib/base/abs_mmio.h"
+#include "sw/device/lib/base/abs_mmio.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
 #include "rstmgr_regs.h"
